+++ java/src/test/java/io/cucumber/query/QueryAcceptanceTest.java
@@ -4,12 +4,13 @@ import com.fasterxml.jackson.core.util.DefaultPrettyPrinter;
 import io.cucumber.messages.Convertor;
 import io.cucumber.messages.NdjsonToMessageIterable;
 import io.cucumber.messages.types.Envelope;
-import io.cucumber.messages.types.Feature;
 import io.cucumber.messages.types.Hook;
 import io.cucumber.messages.types.Pickle;
 import io.cucumber.messages.types.PickleStep;
 import io.cucumber.messages.types.Step;
 import io.cucumber.messages.types.StepDefinition;
+import io.cucumber.messages.types.Suggestion;
+import io.cucumber.messages.types.TestCase;
 import io.cucumber.messages.types.TestCaseFinished;
 import io.cucumber.messages.types.TestCaseStarted;
 import io.cucumber.messages.types.TestStep;
@@ -26,22 +27,17 @@ import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.Comparator;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
-import java.util.stream.Stream;
+import java.util.function.Function;
 
 import static com.fasterxml.jackson.core.util.DefaultIndenter.SYSTEM_LINEFEED_INSTANCE;
 import static io.cucumber.query.Jackson.OBJECT_MAPPER;
-import static io.cucumber.query.NamingStrategy.ExampleName.PICKLE;
-import static io.cucumber.query.NamingStrategy.FeatureName.EXCLUDE;
-import static io.cucumber.query.NamingStrategy.Strategy.LONG;
-import static io.cucumber.query.NamingStrategy.Strategy.SHORT;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Arrays.asList;
 import static java.util.stream.Collectors.toList;
@@ -50,48 +46,56 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class QueryAcceptanceTest {
     private static final NdjsonToMessageIterable.Deserializer deserializer = (json) -> OBJECT_MAPPER.readValue(json, Envelope.class);
 
-    static List<TestCase> acceptance() {
+    static List<QueryTestCase> acceptance() {
+        List<QueryTestCase> testCases = new ArrayList<>();
 
-        return Stream.of(
-                        Paths.get("../testdata/attachments.ndjson"),
-                        Paths.get("../testdata/empty.ndjson"),
-                        Paths.get("../testdata/hooks.ndjson"),
-                        Paths.get("../testdata/minimal.ndjson"),
-                        Paths.get("../testdata/rules.ndjson"),
-                        Paths.get("../testdata/examples-tables.ndjson")
-                )
-                .map(TestCase::new)
-                .sorted(Comparator.comparing(testCase -> testCase.source))
-                .collect(toList());
+        List<Path> sources = getSources();
+        Map<String, Function<Query, Object>> queries = createQueries();
+
+        sources.forEach(source ->
+                queries.forEach((methodName, query) ->
+                        testCases.add(new QueryTestCase(methodName, source, query))));
+
+        return testCases;
+    }
+
+    private static List<Path> getSources() {
+        return Arrays.asList(
+                Paths.get("../testdata/src/attachments.ndjson"),
+                Paths.get("../testdata/src/empty.ndjson"),
+                Paths.get("../testdata/src/hooks.ndjson"),
+                Paths.get("../testdata/src/minimal.ndjson"),
+                Paths.get("../testdata/src/rules.ndjson"),
+                Paths.get("../testdata/src/examples-tables.ndjson")
+        );
     }
 
     @ParameterizedTest
     @MethodSource("acceptance")
-    void test(TestCase testCase) throws IOException {
+    void test(QueryTestCase testCase) throws IOException {
         ByteArrayOutputStream bytes = writeQueryResults(testCase, new ByteArrayOutputStream());
         String expected = new String(Files.readAllBytes(testCase.expected), UTF_8);
         String actual = new String(bytes.toByteArray(), UTF_8);
         assertThat(actual).isEqualTo(expected);
     }
 
-
     @ParameterizedTest
     @MethodSource("acceptance")
     @Disabled
-    void updateExpectedQueryResultFiles(TestCase testCase) throws IOException {
+    void updateExpectedQueryResultFiles(QueryTestCase testCase) throws IOException {
         try (OutputStream out = Files.newOutputStream(testCase.expected)) {
             writeQueryResults(testCase, out);
         }
     }
 
-    private static <T extends OutputStream> T writeQueryResults(TestCase testCase, T out) throws IOException {
+    private static <T extends OutputStream> T writeQueryResults(QueryTestCase testCase, T out) throws IOException {
         try (InputStream in = Files.newInputStream(testCase.source)) {
             try (NdjsonToMessageIterable envelopes = new NdjsonToMessageIterable(in, deserializer)) {
                 Query query = new Query();
                 for (Envelope envelope : envelopes) {
                     query.update(envelope);
                 }
-                Map<String, Object> queryResults = createQueryResults(query);
+                Object queryResults = testCase.query.apply(query);
                 DefaultPrettyPrinter prettyPrinter = new DefaultPrettyPrinter()
                         .withArrayIndenter(SYSTEM_LINEFEED_INSTANCE);
                 OBJECT_MAPPER.writer(prettyPrinter).writeValue(out, queryResults);
@@ -100,166 +104,231 @@ public class QueryAcceptanceTest {
         return out;
     }
 
-    private static Map<String, Object> createQueryResults(Query query) {
+    static Map<String, Function<Query, Object>> createQueries() {
 
-        Map<String, Object> results = new LinkedHashMap<>();
+        Map<String, Function<Query, Object>> queries = new LinkedHashMap<>();
 
-        results.put("countMostSevereTestStepResultStatus", query.countMostSevereTestStepResultStatus());
-        results.put("countTestCasesStarted", query.countTestCasesStarted());
-        results.put("findAllPickles", query.findAllPickles().size());
-        results.put("findAllPickleSteps", query.findAllPickleSteps().size());
-        results.put("findAllTestCaseStarted", query.findAllTestCaseStarted().size());
-        results.put("findAllTestSteps", query.findAllTestSteps().size());
-        results.put("findAllTestCaseStartedGroupedByFeature", query.findAllTestCaseStartedGroupedByFeature()
-                .entrySet()
-                .stream()
-                .map(entry -> Arrays.asList(entry.getKey().map(Feature::getName), entry.getValue().stream()
-                        .map(TestCaseStarted::getId)
-                        .collect(toList()))));
-        results.put("findAttachmentsBy", query.findAllTestCaseStarted().stream()
+        queries.put("countMostSevereTestStepResultStatus", Query::countMostSevereTestStepResultStatus);
+        queries.put("countTestCasesStarted", Query::countTestCasesStarted);
+        queries.put("findAllPickles", (query) -> query.findAllPickles().size());
+        queries.put("findAllPickleSteps", (query) -> query.findAllPickleSteps().size());
+        queries.put("findAllTestCaseStarted", (query) -> query.findAllTestCaseStarted().size());
+        queries.put("findAllTestCaseFinished", (query) -> query.findAllTestCaseFinished().size());
+        queries.put("findAllTestSteps", (query) -> query.findAllTestSteps().size());
+        queries.put("findAllTestStepsStarted", (query) -> query.findAllTestStepStarted().size());
+        queries.put("findAllTestStepsFinished", (query) -> query.findAllTestStepFinished().size());
+        queries.put("findAllTestCases", (query) -> query.findAllTestCases().size());
+        queries.put("findAttachmentsBy", (query) -> query.findAllTestCaseStarted().stream()
                 .map(query::findTestStepFinishedAndTestStepBy)
                 .flatMap(Collection::stream)
                 .map(Map.Entry::getKey)
                 .map(query::findAttachmentsBy)
                 .flatMap(Collection::stream)
-                        .map(attachment -> Arrays.asList(
-                                attachment.getTestStepId(),
-                                attachment.getTestCaseStartedId(),
-                                attachment.getMediaType(),
-                                attachment.getContentEncoding()
-                        ))
+                .map(attachment -> Arrays.asList(
+                        attachment.getTestStepId(),
+                        attachment.getTestCaseStartedId(),
+                        attachment.getMediaType(),
+                        attachment.getContentEncoding()
+                ))
                 .collect(toList()));
-        results.put("findFeatureBy", query.findAllTestCaseStarted().stream()
-                .map(query::findFeatureBy)
-                .map(feature -> feature.map(Feature::getName))
-                .collect(toList()));
-        results.put("findHookBy", query.findAllTestSteps().stream()
+        queries.put("findHookBy", (query) -> query.findAllTestSteps().stream()
                 .map(query::findHookBy)
                 .map(hook -> hook.map(Hook::getId))
                 .filter(Optional::isPresent)
                 .collect(toList()));
-        results.put("findLocationOf", query.findAllPickles().stream()
+        queries.put("findLocationOf", (query) -> query.findAllPickles().stream()
                 .map(query::findLocationOf)
                 .filter(Optional::isPresent)
                 .collect(toList()));
-        results.put("findMeta", query.findMeta().map(meta -> meta.getImplementation().getName()));
-        results.put("findMostSevereTestStepResultBy", query.findAllTestCaseStarted().stream()
-                .map(query::findMostSevereTestStepResultBy)
-                .map(testStepResult -> testStepResult.map(TestStepResult::getStatus))
-                .collect(toList()));
+        queries.put("findMeta", (query) -> query.findMeta().map(meta -> meta.getImplementation().getName()));
 
-        Map<String, Object> names = new LinkedHashMap<>();
-        names.put("long", query.findAllPickles().stream()
-                .map(pickle -> query.findNameOf(pickle, NamingStrategy.strategy(LONG).build()))
-                .collect(toList()));
-        names.put("excludeFeatureName", query.findAllPickles().stream()
-                .map(pickle -> query.findNameOf(pickle, NamingStrategy.strategy(LONG).featureName(EXCLUDE).build()))
-                .collect(toList()));
-        names.put("longPickleName", query.findAllPickles().stream()
-                .map(pickle -> query.findNameOf(pickle, NamingStrategy.strategy(LONG).exampleName(PICKLE).build()))
-                .collect(toList()));
-        names.put("short", query.findAllPickles().stream()
-                .map(pickle -> query.findNameOf(pickle, NamingStrategy.strategy(SHORT).build()))
-                .collect(toList()));
-        names.put("shortPickleName", query.findAllPickles().stream()
-                .map(pickle -> query.findNameOf(pickle, NamingStrategy.strategy(SHORT).exampleName(PICKLE).build()))
-                .collect(toList()));
+        queries.put("findMostSevereTestStepResultBy", (query) -> {
+            Map<String, Object> results = new LinkedHashMap<>();
+            results.put("testCaseStarted", query.findAllTestCaseStarted().stream()
+                    .map(query::findMostSevereTestStepResultBy)
+                    .map(testStepResult -> testStepResult.map(TestStepResult::getStatus))
+                    .filter(Optional::isPresent)
+                    .map(Optional::get)
+                    .collect(toList()));
+            results.put("testCaseFinished", query.findAllTestCaseFinished().stream()
+                    .map(query::findMostSevereTestStepResultBy)
+                    .map(testStepResult -> testStepResult.map(TestStepResult::getStatus))
+                    .filter(Optional::isPresent)
+                    .map(Optional::get)
+                    .collect(toList()));
+            return results;
+        });
 
-        results.put("findNameOf", names);
+        queries.put("findPickleBy", (query) -> {
+            Map<String, Object> results = new LinkedHashMap<>();
+            results.put("testCaseStarted", query.findAllTestCaseStarted().stream()
+                    .map(query::findPickleBy)
+                    .map(pickle -> pickle.map(Pickle::getName))
+                    .collect(toList()));
+            results.put("testCaseFinished", query.findAllTestCaseFinished().stream()
+                    .map(query::findPickleBy)
+                    .map(pickle -> pickle.map(Pickle::getName))
+                    .collect(toList()));
+            results.put("testStepStarted", query.findAllTestStepStarted().stream()
+                    .map(query::findPickleBy)
+                    .map(pickle -> pickle.map(Pickle::getName))
+                    .collect(toList()));
+            results.put("testStepFinished", query.findAllTestStepFinished().stream()
+                    .map(query::findPickleBy)
+                    .map(pickle -> pickle.map(Pickle::getName))
+                    .collect(toList()));
+            return results;
+        });
 
-        results.put("findPickleBy", query.findAllTestCaseStarted().stream()
-                .map(query::findPickleBy)
-                .map(pickle -> pickle.map(Pickle::getName))
-                .collect(toList()));
-        results.put("findPickleStepBy", query.findAllTestSteps().stream()
+        queries.put("findPickleStepBy", (query) -> query.findAllTestSteps().stream()
                 .map(query::findPickleStepBy)
                 .map(pickleStep -> pickleStep.map(PickleStep::getText))
                 .filter(Optional::isPresent)
                 .collect(toList()));
-        results.put("findStepBy", query.findAllPickleSteps().stream()
+        queries.put("findStepBy", (query) -> query.findAllPickleSteps().stream()
                 .map(query::findStepBy)
                 .map(step -> step.map(Step::getText))
                 .collect(toList()));
-        results.put("findStepDefinitionsBy", query.findAllTestSteps().stream()
+        queries.put("findStepDefinitionsBy", (query) -> query.findAllTestSteps().stream()
                 .map(query::findStepDefinitionsBy)
                 .map(stepDefinitions -> stepDefinitions.stream().map(StepDefinition::getId)
                         .collect(toList()))
                 .collect(toList()));
-        results.put("findUnambiguousStepDefinitionBy", query.findAllTestSteps().stream()
+
+        queries.put("findSuggestionsBy", (query) -> {
+            Map<String, Object> results = new LinkedHashMap<>();
+            results.put("pickleStep", query.findAllPickleSteps().stream()
+                    .map(query::findSuggestionsBy)
+                    .flatMap(Collection::stream)
+                    .map(Suggestion::getId)
+                    .collect(toList()));
+            results.put("pickle", query.findAllPickles().stream()
+                    .map(query::findSuggestionsBy)
+                    .flatMap(Collection::stream)
+                    .map(Suggestion::getId)
+                    .collect(toList()));
+            return results;
+        });
+
+        queries.put("findUnambiguousStepDefinitionBy", (query) -> query.findAllTestSteps().stream()
                 .map(query::findUnambiguousStepDefinitionBy)
                 .filter(Optional::isPresent)
                 .map(Optional::get)
                 .map(StepDefinition::getId));
-        results.put("findTestCaseBy", query.findAllTestCaseStarted().stream()
-                .map(query::findTestCaseBy)
-                .map(testCase -> testCase.map(io.cucumber.messages.types.TestCase::getId))
-                .collect(toList()));
-        results.put("findTestCaseDurationBy", query.findAllTestCaseStarted().stream()
-                .map(query::findTestCaseDurationBy)
-                .map(duration -> duration.map(Convertor::toMessage))
-                .collect(toList()));
-        results.put("findTestCaseFinishedBy", query.findAllTestCaseStarted().stream()
+
+        queries.put("findTestCaseStartedBy", (query) -> {
+            Map<String, Object> results = new LinkedHashMap<>();
+            results.put("testStepStarted", query.findAllTestStepStarted().stream()
+                    .map(query::findTestCaseStartedBy)
+                    .map(testCase -> testCase.map(TestCaseStarted::getId))
+                    .collect(toList()));
+            results.put("testStepFinished", query.findAllTestStepFinished().stream()
+                    .map(query::findTestCaseStartedBy)
+                    .map(testCase -> testCase.map(TestCaseStarted::getId))
+                    .collect(toList()));
+            return results;
+        });
+
+        queries.put("findTestCaseBy", (query) -> {
+            Map<String, Object> results = new LinkedHashMap<>();
+            results.put("testCaseStarted", query.findAllTestCaseStarted().stream()
+                    .map(query::findTestCaseBy)
+                    .map(testCase -> testCase.map(TestCase::getId))
+                    .collect(toList()));
+            results.put("testCaseFinished", query.findAllTestCaseFinished().stream()
+                    .map(query::findTestCaseBy)
+                    .map(testCase -> testCase.map(TestCase::getId))
+                    .collect(toList()));
+            results.put("testStepStarted", query.findAllTestStepStarted().stream()
+                    .map(query::findTestCaseBy)
+                    .map(testCase -> testCase.map(TestCase::getId))
+                    .collect(toList()));
+            results.put("testStepFinished", query.findAllTestStepFinished().stream()
+                    .map(query::findTestCaseBy)
+                    .map(testCase -> testCase.map(TestCase::getId))
+                    .collect(toList()));
+            return results;
+        });
+
+        queries.put("findTestCaseDurationBy", (query) -> {
+            Map<String, Object> results = new LinkedHashMap<>();
+            results.put("testCaseStarted", query.findAllTestCaseStarted().stream()
+                    .map(query::findTestCaseDurationBy)
+                    .map(duration -> duration.map(Convertor::toMessage))
+                    .collect(toList()));
+            results.put("testCaseFinished", query.findAllTestCaseFinished().stream()
+                    .map(query::findTestCaseDurationBy)
+                    .map(duration -> duration.map(Convertor::toMessage))
+                    .collect(toList()));
+            return results;
+        });
+
+        queries.put("findTestCaseFinishedBy", (query) -> query.findAllTestCaseStarted().stream()
                 .map(query::findTestCaseFinishedBy)
                 .map(testCaseFinished -> testCaseFinished.map(TestCaseFinished::getTestCaseStartedId))
                 .collect(toList()));
-        results.put("findTestRunDuration", query.findTestRunDuration()
+        queries.put("findTestRunDuration", (query) -> query.findTestRunDuration()
                 .map(Convertor::toMessage));
-        results.put("findTestRunFinished", query.findTestRunFinished());
-        results.put("findTestRunStarted", query.findTestRunStarted());
-        results.put("findTestStepByTestStepStarted", query.findAllTestCaseStarted().stream()
+        queries.put("findTestRunFinished", Query::findTestRunFinished);
+        queries.put("findTestRunStarted", Query::findTestRunStarted);
+        queries.put("findTestStepByTestStepStarted", (query) -> query.findAllTestCaseStarted().stream()
                 .map(query::findTestStepsStartedBy)
                 .flatMap(Collection::stream)
                 .map(query::findTestStepBy)
                 .map(testStep -> testStep.map(TestStep::getId))
                 .collect(toList()));
-        results.put("findTestStepByTestStepFinished", query.findAllTestCaseStarted().stream()
-                .map(query::findTestStepsFinishedBy)
-                .flatMap(Collection::stream)
-                .map(query::findTestStepBy)
-                .map(testStep -> testStep.map(TestStep::getId))
-                .collect(toList()));
-        results.put("findTestStepsFinishedBy", query.findAllTestCaseStarted().stream()
+
+        queries.put("findTestStepByTestStepFinished", (query) -> {
+            Map<String, Object> results = new LinkedHashMap<>();
+
+            results.put("testCaseStarted", query.findAllTestCaseStarted().stream()
+                    .map(query::findTestStepsFinishedBy)
+                    .flatMap(Collection::stream)
+                    .map(query::findTestStepBy)
+                    .map(testStep -> testStep.map(TestStep::getId))
+                    .collect(toList()));
+            results.put("testCaseFinished", query.findAllTestCaseFinished().stream()
+                    .map(query::findTestStepsFinishedBy)
+                    .flatMap(Collection::stream)
+                    .map(query::findTestStepBy)
+                    .map(testStep -> testStep.map(TestStep::getId))
+                    .collect(toList()));
+
+            return results;
+        });
+
+        queries.put("findTestStepsFinishedBy", (query) -> query.findAllTestCaseStarted().stream()
                 .map(query::findTestStepsFinishedBy)
                 .map(testStepFinisheds -> testStepFinisheds.stream().map(TestStepFinished::getTestStepId).collect(toList()))
                 .collect(toList()));
-        results.put("findTestStepFinishedAndTestStepBy", query.findAllTestCaseStarted().stream()
+        queries.put("findTestStepFinishedAndTestStepBy", (query) -> query.findAllTestCaseStarted().stream()
                 .map(query::findTestStepFinishedAndTestStepBy)
                 .flatMap(Collection::stream)
                 .map(entry -> asList(entry.getKey().getTestStepId(), entry.getValue().getId()))
                 .collect(toList()));
 
-        return results;
+        return queries;
     }
 
-    static class TestCase {
+    static class QueryTestCase {
+        private final String methodName;
+        private final String name;
         private final Path source;
         private final Path expected;
+        private final Function<Query, Object> query;
 
-        private final String name;
-
-        TestCase(Path source) {
+        QueryTestCase(String methodName, Path source, Function<Query, Object> query) {
+            this.methodName = methodName;
             this.source = source;
+            this.query = query;
             String fileName = source.getFileName().toString();
             this.name = fileName.substring(0, fileName.lastIndexOf(".ndjson"));
-            this.expected = source.getParent().resolve(name + ".query-results.json");
+            this.expected = source.getParent().resolve(name + "." + methodName + ".results.json");
         }
 
         @Override
         public String toString() {
-            return name;
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) return true;
-            if (o == null || getClass() != o.getClass()) return false;
-            TestCase testCase = (TestCase) o;
-            return source.equals(testCase.source);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(source);
+            return name + " -> " + methodName;
         }
     }