+++ CHANGELOG.md
@@ -6,9 +6,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-### Changed
-- [Java] increase dependency messages to at least v28.1 ([#87](https://github.com/cucumber/query/pull/87))
+### Added
+- Update dependency `messages` up to v29 ([#101](https://github.com/cucumber/query/pull/101))
+- Added more queries to find messages by `TestCaseFinished` and `TestStepFinished` ([#77](https://github.com/cucumber/query/pull/77))
+- Added queries to find suggestions by `Pickle` and `PickleStep` ([#102](https://github.com/cucumber/query/pull/102))
 
+### Deprecated
+- Deprecated various lineage derived methods ([#84](https://github.com/cucumber/query/pull/84))
+
+### Removed
+- [JavaScript] Remove Node.js 18 support ([#84](https://github.com/cucumber/query/pull/84))
+- Removed support for `messages` below v29
 
 ## [13.6.0] - 2025-08-11
 ### Changed